refactor: Gifthub controller 코드 포맷 일치
* 가독성을 위해 다른 controller와 동일하게 코드 포맷 일치
* 사용하지 않는 dto 삭제
