+++ src/main/java/practice/fundingboost2/item/gifthub/app/dto/GetGifthubResponseDto.java
@@ -1,0 +1,10 @@
+package practice.fundingboost2.item.gifthub.app.dto;
+
+public record GetGifthubResponseDto(
+    Long itemId,
+    String itemName,
+    String itemImage,
+    String option,
+    Integer quantity) {
+
+}
