--- src/main/java/practice/fundingboost2/item/gifthub/app/dto/UpdateGifthubRequestDto.java
@@ -1,5 +1,0 @@
-package practice.fundingboost2.item.gifthub.app.dto;
-
-public record UpdateGifthubRequestDto(Integer quantity) {
-
-}
