+++ src/main/java/practice/fundingboost2/item/gifthub/repo/GifthubQueryRepository.java
@@ -2,7 +2,7 @@ package practice.fundingboost2.item.gifthub.repo;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import practice.fundingboost2.item.gifthub.ui.dto.GetGifthubResponseDto;
+import practice.fundingboost2.item.gifthub.app.dto.GetGifthubResponseDto;
 
 public interface GifthubQueryRepository {
