+++ src/main/java/practice/fundingboost2/item/gifthub/ui/GifthubController.java
@@ -21,20 +21,44 @@ public class GifthubController {
     private final GifthubService gifthubService;
 
     @PostMapping("/{item_id}/{option_id}")
-    public ResponseDto<CommonSuccessDto> addItemToCart(@Auth Long memberId, @PathVariable("item_id") Long itemId,
-        @PathVariable("option_id") Long optionId) {
+    public ResponseDto<CommonSuccessDto> addItemToCart(
+        @Auth
+        Long memberId,
+
+        @PathVariable("item_id")
+        Long itemId,
+
+        @PathVariable("option_id")
+        Long optionId) {
         return ResponseDto.ok(gifthubService.addToCart(memberId, itemId, optionId));
     }
 
     @DeleteMapping("{item_id}/{option_id}")
-    public ResponseDto<CommonSuccessDto> deleteItemFromCart(@Auth Long memberId, @PathVariable("item_id") Long itemId,
-        @PathVariable("option_id") Long optionId) {
+    public ResponseDto<CommonSuccessDto> deleteItemFromCart(
+        @Auth
+        Long memberId,
+
+        @PathVariable("item_id")
+        Long itemId,
+
+        @PathVariable("option_id")
+        Long optionId) {
         return ResponseDto.ok(gifthubService.deleteFromCart(memberId, itemId, optionId));
     }
 
     @PatchMapping("/{item_id}/{option_id}")
-    public ResponseDto<CommonSuccessDto> updateItemFromCart(@Auth Long memberId, @PathVariable("item_id") Long itemId,
-        @PathVariable("option_id") Long optionId, @RequestParam("quantity") Integer quantity) {
+    public ResponseDto<CommonSuccessDto> updateItemFromCart(
+        @Auth
+        Long memberId,
+
+        @PathVariable("item_id")
+        Long itemId,
+
+        @PathVariable("option_id")
+        Long optionId,
+
+        @RequestParam("quantity")
+        Integer quantity) {
         return ResponseDto.ok(gifthubService.updateCart(memberId, itemId, optionId, quantity));
     }
 }
