+++ src/main/java/practice/fundingboost2/item/gifthub/ui/GifthubQueryController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 import practice.fundingboost2.common.dto.ResponseDto;
 import practice.fundingboost2.config.security.annotation.Auth;
 import practice.fundingboost2.item.gifthub.repo.GifthubQueryRepository;
-import practice.fundingboost2.item.gifthub.ui.dto.GetGifthubResponseDto;
+import practice.fundingboost2.item.gifthub.app.dto.GetGifthubResponseDto;
 
 @RestController
 @RequestMapping("/api/gifthub")
@@ -19,7 +19,11 @@ public class GifthubQueryController {
     private final GifthubQueryRepository gifthubQueryRepository;
 
     @GetMapping
-    public ResponseDto<Page<GetGifthubResponseDto>> getGifthub(@Auth Long memberId, Pageable pageable) {
+    public ResponseDto<Page<GetGifthubResponseDto>> getGifthub(
+        @Auth
+        Long memberId,
+
+        Pageable pageable) {
         return ResponseDto.ok(gifthubQueryRepository.getAllGifthub(memberId, pageable));
     }
 }
