--- src/main/java/practice/fundingboost2/item/gifthub/ui/dto/GetGifthubListResponseDto.java
@@ -1,10 +1,0 @@
-package practice.fundingboost2.item.gifthub.ui.dto;
-
-import java.util.List;
-
-public record GetGifthubListResponseDto(List<GetGifthubResponseDto> gifthubs) {
-
-    public static GetGifthubListResponseDto from(List<GetGifthubResponseDto> gifthubs) {
-        return new GetGifthubListResponseDto(gifthubs);
-    }
-}
