--- src/main/java/practice/fundingboost2/item/gifthub/ui/dto/GetGifthubResponseDto.java
@@ -1,5 +1,0 @@
-package practice.fundingboost2.item.gifthub.ui.dto;
-
-public record GetGifthubResponseDto(Long itemId, String itemName, String itemImage, String option, Integer quantity) {
-
-}
